Added decrease button and fixed markup
Added a "decrease" button to the cart which allows a user to remove a single item from their cart. If it's the last item then it deletes the cart altogether. Fixed some broken markup causing forms to be moved into the "#side" section of my CSS structure instead of "#main".